refactor: split TUIModels, retry, yamlParser, logger into focused sub-modules (#53)
Split four large files (448-566 LOC) into focused sub-modules, each under 300 LOC,
while preserving full backward compatibility via re-export shims.

- src/models/tui/: TUIConfig, TUISession, TUIResults (split from TUIModels.ts 566 LOC)
- src/utils/retry/: types, RetryExecutor, CircuitBreaker (split from retry.ts 551 LOC)
- src/utils/yaml/: types, YamlLoader, YamlVariableSubstitution, YamlValidator (split from yamlParser.ts 532 LOC)
- src/utils/logging/: LogFormatter, LogTransport (split from logger.ts 448 LOC)

All original entry-point files now re-export from sub-modules (~10-234 LOC).
tsc --noEmit passes; retry.test.ts (30/30) and yamlParser.test.ts (28/28) pass.

Closes #53
