+++ src/models/tui/index.ts
@@ -1,0 +1,7 @@
+/**
+ * TUI sub-module barrel exports
+ */
+
+export * from './TUIConfig';
+export * from './TUISession';
+export * from './TUIResults';
