+++ src/utils/logger.ts
@@ -1,79 +1,23 @@
 /**
  * Winston-based logging system for the Agentic Testing System
  * Provides configurable logging with different levels, formats, and outputs
+ *
+ * This file re-exports types from logging/ sub-modules and provides the
+ * TestLogger class plus the singleton convenience object.
  */
 
 import winston from 'winston';
-import path from 'path';
-import fs from 'fs';
+import {
+  LogLevel,
+  LogContext,
+  LoggerConfig,
+  DEFAULT_LOGGER_CONFIG,
+  buildTransports
+} from './logging/LogTransport';
 
-/**
- * Log levels enumeration
- */
-export enum LogLevel {
-  ERROR = 'error',
-  WARN = 'warn',
-  INFO = 'info',
-  HTTP = 'http',
-  DEBUG = 'debug'
-}
-
-/**
- * Log context interface for structured logging
- */
-export interface LogContext {
-  /** Test scenario ID */
-  scenarioId?: string;
-  /** Test step index */
-  stepIndex?: number;
-  /** Component or module name */
-  component?: string;
-  /** Session or run ID */
-  sessionId?: string;
-  /** Additional metadata */
-  metadata?: Record<string, unknown>;
-}
-
-/**
- * Logger configuration options
- */
-export interface LoggerConfig {
-  /** Log level threshold */
-  level: LogLevel;
-  /** Output directory for log files */
-  logDir: string;
-  /** Whether to log to console */
-  enableConsole: boolean;
-  /** Whether to log to file */
-  enableFile: boolean;
-  /** Maximum size of each log file in bytes */
-  maxFileSize: number;
-  /** Maximum number of log files to keep */
-  maxFiles: number;
-  /** Whether to compress rotated log files */
-  compress: boolean;
-  /** Custom log format */
-  format?: string;
-  /** Whether to include timestamps */
-  includeTimestamp: boolean;
-  /** Whether to include stack traces for errors */
-  includeStackTrace: boolean;
-}
-
-/**
- * Default logger configuration
- */
-const DEFAULT_CONFIG: LoggerConfig = {
-  level: LogLevel.INFO,
-  logDir: './logs',
-  enableConsole: true,
-  enableFile: true,
-  maxFileSize: 10 * 1024 * 1024, // 10MB
-  maxFiles: 5,
-  compress: true,
-  includeTimestamp: true,
-  includeStackTrace: true
-};
+// Re-export so existing imports from './logger' continue to work
+export { LogLevel };
+export type { LogContext, LoggerConfig };
 
 /**
  * Enhanced Winston logger with testing-specific features
@@ -84,101 +28,12 @@ export class TestLogger {
   private context: LogContext = {};
 
   constructor(config: Partial<LoggerConfig> = {}) {
-    this.config = { ...DEFAULT_CONFIG, ...config };
+    this.config = { ...DEFAULT_LOGGER_CONFIG, ...config };
     this.logger = this.createLogger();
   }
 
-  /**
-   * Create the Winston logger instance with configured transports
-   */
   private createLogger(): winston.Logger {
-    // Ensure log directory exists
-    if (this.config.enableFile && !fs.existsSync(this.config.logDir)) {
-      fs.mkdirSync(this.config.logDir, { recursive: true });
-    }
-
-    const transports: winston.transport[] = [];
-
-    // Console transport
-    if (this.config.enableConsole) {
-      transports.push(
-        new winston.transports.Console({
-          level: this.config.level,
-          format: winston.format.combine(
-            winston.format.colorize({ all: true }),
-            winston.format.timestamp({ format: 'YYYY-MM-DD HH:mm:ss' }),
-            winston.format.printf(({ timestamp, level, message, ...meta }) => {
-              let output = `${timestamp} [${level}]`;
-              
-              // Add context information
-              if (this.context.scenarioId) {
-                output += ` [${this.context.scenarioId}]`;
-              }
-              if (this.context.component) {
-                output += ` [${this.context.component}]`;
-              }
-              
-              output += `: ${message}`;
-              
-              // Add metadata if present
-              if (Object.keys(meta).length > 0) {
-                output += ` ${JSON.stringify(meta)}`;
-              }
-              
-              return output;
-            })
-          )
-        })
-      );
-    }
-
-    // File transports
-    if (this.config.enableFile) {
-      // Combined log file
-      transports.push(
-        new winston.transports.File({
-          filename: path.join(this.config.logDir, 'combined.log'),
-          level: this.config.level,
-          maxsize: this.config.maxFileSize,
-          maxFiles: this.config.maxFiles,
-          format: winston.format.combine(
-            winston.format.timestamp(),
-            winston.format.errors({ stack: this.config.includeStackTrace }),
-            winston.format.json()
-          )
-        })
-      );
-
-      // Error-only log file
-      transports.push(
-        new winston.transports.File({
-          filename: path.join(this.config.logDir, 'error.log'),
-          level: LogLevel.ERROR,
-          maxsize: this.config.maxFileSize,
-          maxFiles: this.config.maxFiles,
-          format: winston.format.combine(
-            winston.format.timestamp(),
-            winston.format.errors({ stack: true }),
-            winston.format.json()
-          )
-        })
-      );
-
-      // Test-specific log file
-      transports.push(
-        new winston.transports.File({
-          filename: path.join(this.config.logDir, 'test-execution.log'),
-          level: this.config.level,
-          maxsize: this.config.maxFileSize,
-          maxFiles: this.config.maxFiles,
-          format: winston.format.combine(
-            winston.format.timestamp(),
-            winston.format.errors({ stack: this.config.includeStackTrace }),
-            winston.format.json()
-          )
-        })
-      );
-    }
+    const transports = buildTransports(this.config, () => this.context);
 
     return winston.createLogger({
       level: this.config.level,
@@ -192,23 +47,14 @@ export class TestLogger {
     });
   }
 
-  /**
-   * Set the logging context for structured logging
-   */
   setContext(context: LogContext): void {
     this.context = { ...this.context, ...context };
   }
 
-  /**
-   * Clear the current logging context
-   */
   clearContext(): void {
     this.context = {};
   }
 
-  /**
-   * Get the current logging context
-   */
   getContext(): LogContext {
     return { ...this.context };
   }
@@ -248,34 +94,25 @@ export class TestLogger {
     this.logger.debug(message, { ...this.context, ...(typeof meta === 'object' && meta !== null ? meta as Record<string, unknown> : { meta }) });
   }
 
-  /**
-   * Log test scenario start
-   */
   scenarioStart(scenarioId: string, scenarioName: string): void {
     this.setContext({ scenarioId });
-    this.info(`Starting test scenario: ${scenarioName}`, { 
+    this.info(`Starting test scenario: ${scenarioName}`, {
       event: 'scenario_start',
       scenarioId,
-      scenarioName 
+      scenarioName
     });
   }
 
-  /**
-   * Log test scenario completion
-   */
   scenarioEnd(scenarioId: string, status: string, duration: number): void {
-    this.info(`Test scenario completed: ${status}`, { 
+    this.info(`Test scenario completed: ${status}`, {
       event: 'scenario_end',
       scenarioId,
       status,
-      duration 
+      duration
     });
     this.clearContext();
   }
 
-  /**
-   * Log test step execution
-   */
   stepExecution(stepIndex: number, action: string, target: string): void {
     this.debug(`Executing step ${stepIndex + 1}: ${action}`, {
       event: 'step_execution',
@@ -285,9 +122,6 @@ export class TestLogger {
     });
   }
 
-  /**
-   * Log test step completion
-   */
   stepComplete(stepIndex: number, status: string, duration: number): void {
     this.debug(`Step ${stepIndex + 1} completed: ${status}`, {
       event: 'step_complete',
@@ -309,9 +143,6 @@ export class TestLogger {
     });
   }
 
-  /**
-   * Log screenshot capture
-   */
   screenshot(filename: string, stepIndex?: number): void {
     this.debug(`Screenshot captured: ${filename}`, {
       event: 'screenshot',
@@ -320,9 +151,6 @@ export class TestLogger {
     });
   }
 
-  /**
-   * Log command execution
-   */
   commandExecution(command: string, workingDir?: string): void {
     this.debug(`Executing command: ${command}`, {
       event: 'command_execution',
@@ -331,9 +159,6 @@ export class TestLogger {
     });
   }
 
-  /**
-   * Log command completion
-   */
   commandComplete(command: string, exitCode: number, duration: number): void {
     const level = exitCode === 0 ? 'info' : 'warn';
     this.logger[level](`Command completed with exit code ${exitCode}`, {
@@ -344,43 +169,29 @@ export class TestLogger {
     });
   }
 
-  /**
-   * Create a child logger with additional context
-   */
   child(context: LogContext): TestLogger {
     const childLogger = new TestLogger(this.config);
     childLogger.setContext({ ...this.context, ...context });
     return childLogger;
   }
 
-  /**
-   * Change the log level at runtime
-   */
   setLevel(level: LogLevel): void {
     this.config.level = level;
     this.logger.level = level;
-    
-    // Update all transports
+
     this.logger.transports.forEach(transport => {
       transport.level = level;
     });
   }
 
-  /**
-   * Get current log level
-   */
   getLevel(): LogLevel {
     return this.config.level;
   }
 
-  /**
-   * Flush all log transports
-   */
   async flush(): Promise<void> {
     return new Promise((resolve) => {
-      // Winston doesn't have a built-in flush method, so we use a workaround
       let pendingTransports = this.logger.transports.length;
-      
+
       if (pendingTransports === 0) {
         resolve();
         return;
@@ -388,7 +199,6 @@ export class TestLogger {
 
       this.logger.transports.forEach((transport) => {
         if ('close' in transport && typeof transport.close === 'function') {
-          // Winston v3 close() is synchronous (no callback)
           transport.close();
         }
         pendingTransports--;
@@ -399,9 +209,6 @@ export class TestLogger {
     });
   }
 
-  /**
-   * Close the logger and clean up resources
-   */
   async close(): Promise<void> {
     await this.flush();
     this.logger.close();
@@ -416,8 +223,7 @@ export function createLogger(config?: Partial<LoggerConfig>): TestLogger {
 }
 
 /**
- * Mutable active logger — swapped by setupLogger() without mutating any instance.
- * All code should use the `logger` convenience object rather than this directly.
+ * Mutable active logger swapped by setupLogger() without mutating any instance.
  */
 let _activeLogger: TestLogger = new TestLogger();
 
@@ -426,7 +232,6 @@ export const defaultLogger = _activeLogger;
 
 /**
  * Convenience methods that always delegate to the current active logger.
- * Safe to use before or after setupLogger() is called.
  */
 export const logger = {
   error: (message: string, meta?: unknown) => _activeLogger.error(message, meta),
@@ -440,8 +245,7 @@ export const logger = {
 };
 
 /**
- * Reconfigure the active logger. Replaces the logger instance rather than
- * mutating it, so all subsequent calls through `logger` use the new config.
+ * Reconfigure the active logger. Replaces the logger instance rather than mutating it.
  */
 export function setupLogger(config?: Partial<LoggerConfig>): TestLogger {
   _activeLogger = createLogger(config);
