+++ src/utils/logging/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Logging sub-module barrel exports
+ */
+
+export * from './LogTransport';
+export * from './LogFormatter';
